+++ lib/openapi_parser/schema_validator/string_validator.rb
@@ -16,11 +16,6 @@ class OpenAPIParser::SchemaValidator
       value, err = pattern_validate(value, schema)
       return [nil, err] if err
 
-      unless @datetime_coerce_class.nil?
-        value, err = coerce_date_time(value, schema)
-        return [nil, err] if err
-      end
-
       value, err = validate_max_min_length(value, schema)
       return [nil, err] if err
 
@@ -30,27 +25,16 @@ class OpenAPIParser::SchemaValidator
       value, err = validate_uuid_format(value, schema)
       return [nil, err] if err
 
+      value, err = validate_date_format(value, schema)
+      return [nil, err] if err
+
+      value, err = validate_datetime_format(value, schema)
+      return [nil, err] if err
+
       [value, nil]
     end
 
     private
-
-      # @param [OpenAPIParser::Schemas::Schema] schema
-      def coerce_date_time(value, schema)
-        return parse_date_time(value, schema) if schema.format == 'date-time'
-
-        [value, nil]
-      end
-
-      def parse_date_time(value, schema)
-        begin
-          return @datetime_coerce_class.parse(value), nil
-        rescue ArgumentError => e
-          raise e unless e.message =~ /invalid date/
-        end
-
-        OpenAPIParser::ValidateError.build_error_result(value, schema)
-      end
 
       # @param [OpenAPIParser::Schemas::Schema] schema
       def pattern_validate(value, schema)
@@ -85,6 +69,40 @@ class OpenAPIParser::SchemaValidator
         return [value, nil] if value.match(/[0-9a-fA-F]{8}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{12}/)
 
         return [nil, OpenAPIParser::InvalidUUIDFormat.new(value, schema.object_reference)]
+      end
+
+      def validate_date_format(value, schema)
+        return [value, nil] unless schema.format == 'date'
+
+        begin
+          Date.strptime(value, "%Y-%m-%d")
+        rescue ArgumentError
+          return [nil, OpenAPIParser::InvalidDateFormat.new(value, schema.object_reference)]
+        end
+
+        return [value, nil]
+      end
+
+      def validate_datetime_format(value, schema)
+        return [value, nil] unless schema.format == 'date-time'
+
+        begin
+          if @datetime_coerce_class.nil?
+            # validate only
+            DateTime.rfc3339(value)
+            [value, nil]
+          else
+            # validate and coerce
+            if @datetime_coerce_class == Time
+              [DateTime.rfc3339(value).to_time, nil]
+            else
+              [@datetime_coerce_class.rfc3339(value), nil]
+            end
+          end
+        rescue ArgumentError
+          # when rfc3339(value) failed
+          [nil, OpenAPIParser::InvalidDateTimeFormat.new(value, schema.object_reference)]
+        end
       end
   end
 end