+++ spec/spec_helper.rb
@@ -1,8 +1,10 @@
 require 'bundler'
-require 'yaml'
-require 'pry'
+if Gem::Version.create(RUBY_VERSION) < Gem::Version.create("3.2.0")
+  require 'pry'
+end
 require 'rspec-parameterized'
 require 'simplecov'
+require 'psych'
 
 SimpleCov.start do
   add_filter 'spec'
@@ -24,12 +26,20 @@ RSpec.configure do |config|
   end
 end
 
+def load_yaml_file(path)
+  Psych.safe_load(open(path).read, permitted_classes: [Date, Time])
+end
+
 def normal_schema
-  YAML.load_file('./spec/data/normal.yml')
+  load_yaml_file('./spec/data/normal.yml')
+end
+
+def broken_reference_schema
+  load_yaml_file('./spec/data/reference-broken.yaml')
 end
 
 def petstore_schema
-  YAML.load_file(petstore_schema_path)
+  load_yaml_file(petstore_schema_path)
 end
 
 def petstore_schema_path
@@ -37,7 +47,15 @@ def petstore_schema_path
 end
 
 def petstore_with_discriminator_schema
-  YAML.load_file('./spec/data/petstore-with-discriminator.yaml')
+  load_yaml_file('./spec/data/petstore-with-discriminator.yaml')
+end
+
+def petstore_with_mapped_polymorphism_schema
+  load_yaml_file('./spec/data/petstore-with-mapped-polymorphism.yaml')
+end
+
+def petstore_with_polymorphism_schema
+  load_yaml_file('./spec/data/petstore-with-polymorphism.yaml')
 end
 
 def json_petstore_schema_path
@@ -52,12 +70,16 @@ def yaml_with_unsupported_extension_petstore_schema_path
   './spec/data/petstore.yaml.unsupported_extension'
 end
 
+def path_item_ref_schema_path
+  './spec/data/path-item-ref.yaml'
+end
+
 def path_item_ref_schema
-  YAML.load_file('./spec/data/path-item-ref.yaml')
+  load_yaml_file(path_item_ref_schema_path)
 end
 
 def build_validate_test_schema(new_properties)
-  b = YAML.load_file('./spec/data/validate_test.yaml')
+  b = load_yaml_file('./spec/data/validate_test.yaml')
   obj = b['paths']['/validate_test']['post']['requestBody']['content']['application/json']['schema']['properties']
   obj.merge!(change_string_key(new_properties))
   b